# Performance improvements, up to 20x faster