@Bodigrim 2023-02-18
> What I wonder is whether some of this can be backported to the string-only filepath or whether the new API introduced performance regressions (a few functions were slightly rewritten).

The changes improve performance both for the legacy API and for the new one.

> I get compilation error on 9.2.6:

It's a warning for me, and present in the `master` branch as well.