@hasufell 2023-02-18
If you rebase against master and then push to this repo instead of your fork, we can benefit from the fixed CI.