@hasufell 2023-02-24
@Bodigrim we have a regression:

With this patch (1.4.100.2):

```
> W.splitFileName  "\\\\?\\A:\\fred"
("\\\\?\\A:\\fred", "")
```

With 1.4.100.0:

```
> W.splitFileName  "\\\\?\\A:\\fred"
("\\\\?\\A:\\", "fred")
```

This was by chance found by a property test (they are rather unreliable): https://github.com/haskell/filepath/actions/runs/4263893204/jobs/7421218609#step:5:1112

However, `W.splitDrive` match.