@sternenseemann 2024-01-07
@hasufell That regression was fixed in https://github.com/haskell/filepath/commit/6a1c98dc5eb8021107d59903bbbb60137a6f9dbd, correct?

I noticed at least one other change in behavior which I discovered by chance in the `filepath-bytestring` equivalency test suite. It appears e.g. when comparing `filepath-1.4.100.4` (GHC 9.8.1) to `filepath-1.4.2.2` (GHC 9.4.8):

```console
$ ghci-9.4.8
GHCi, version 9.4.8: https://www.haskell.org/ghc/  :? for help
ghci> import qualified System.FilePath.Windows as W
ghci> W.takeDirectory "//?/a:"
"//?/a:"
$ ghci-9.8.1
GHCi, version 9.8.1: https://www.haskell.org/ghc/  :? for help
ghci> W.takeDirectory "//?/a:"
"//?/"
```

Can you comment whether this is also to be considered a regression?